+++ shillelagh/src/main/java/shillelagh/Column.java
@@ -24,9 +24,9 @@ import static java.lang.annotation.RetentionPolicy.SOURCE;
 
 /** All Database fields must be marked with this annotation. */
 @Retention(SOURCE) @Target(FIELD)
-public @interface Field {
+public @interface Column {
   /** Set the name of the column that shillelagh will use to read and write to the database as. */
-  String columnName() default "";
+  String name() default "";
 
   /** Tells if the field should be converted to a byte array and stored as a blob */
   boolean isBlob() default false;